Remove `CrayExternal` and revert Cray packages to use `Package` base class
This adds repetition, but well...
