+++ repos/spack_repo/builtin/packages/cray_fftw/package.py
@@ -1,12 +1,10 @@
 # Copyright Spack Project Developers. See COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-from spack_repo.builtin.build_systems.cray import CrayExternal
-
 from spack.package import *
 
 
-class CrayFftw(CrayExternal):
+class CrayFftw(Package):
     """FFTW is a C subroutine library for computing the discrete Fourier
     transform (DFT) in one or more dimensions, of arbitrary input
     size, and of both real and complex data (as well as of even/odd
@@ -39,6 +37,21 @@ class CrayFftw(CrayExternal):
     variant("openmp", default=False, description="Enable OpenMP support.")
     variant("mpi", default=True, description="Activate MPI support")
     depends_on("mpi", when="+mpi")
+
+    has_code = False  # Skip attempts to fetch a source that is not available
+
+    # Allows attaching compilers to externals in packages.yaml
+    depends_on("c", type="build")
+
+    requires("platform=linux", msg="Cray software is only available on linux")
+
+    def install(self, spec, prefix):
+        raise InstallError(
+            self.spec.format(
+                "{name} is not installable, you need to specify "
+                "it as an external package in packages.yaml"
+            )
+        )
 
     @property
     def libs(self):
