+++ repos/spack_repo/builtin/packages/cray_libsci/package.py
@@ -1,12 +1,10 @@
 # Copyright Spack Project Developers. See COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-from spack_repo.builtin.build_systems.cray import CrayExternal
-
 from spack.package import *
 
 
-class CrayLibsci(CrayExternal):
+class CrayLibsci(Package):
     """The Cray Scientific Libraries package, LibSci, is a collection of
     numerical routines optimized for best performance on Cray systems."""
 
@@ -43,6 +41,21 @@ class CrayLibsci(CrayExternal):
         "nvhpc": "NVIDIA",
         "rocmcc": "AMD",
     }
+
+    has_code = False  # Skip attempts to fetch a source that is not available
+
+    # Allows attaching compilers to externals in packages.yaml
+    depends_on("c", type="build")
+
+    requires("platform=linux", msg="Cray software is only available on linux")
+
+    def install(self, spec, prefix):
+        raise InstallError(
+            self.spec.format(
+                "{name} is not installable, you need to specify "
+                "it as an external package in packages.yaml"
+            )
+        )
 
     @property
     def modname(self):
