+++ repos/spack_repo/builtin/packages/cray_pmi/package.py
@@ -1,13 +1,10 @@
 # Copyright Spack Project Developers. See COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-
-from spack_repo.builtin.build_systems.cray import CrayExternal
-
 from spack.package import *
 
 
-class CrayPmi(CrayExternal):
+class CrayPmi(Package):
     """Cray's Process Management Interface library"""
 
     homepage = "https://docs.nersc.gov/development/compilers/wrappers/"
@@ -17,6 +14,21 @@ class CrayPmi(CrayExternal):
     version("5.0.17")
     version("5.0.16")
     version("5.0.11")
+
+    has_code = False  # Skip attempts to fetch a source that is not available
+
+    # Allows attaching compilers to externals in packages.yaml
+    depends_on("c", type="build")
+
+    requires("platform=linux", msg="Cray software is only available on linux")
+
+    def install(self, spec, prefix):
+        raise InstallError(
+            self.spec.format(
+                "{name} is not installable, you need to specify "
+                "it as an external package in packages.yaml"
+            )
+        )
 
     @property
     def headers(self):
